Fixed bugs

+++ utils/firebaseConfig.ts
@@ -1,5 +1,5 @@
 import { initializeApp } from "firebase/app";
-import { getAuth, GoogleAuthProvider , setPersistence, browserLocalPersistence} from "firebase/auth";
+import { getAuth, GoogleAuthProvider , setPersistence, browserLocalPersistence , onAuthStateChanged, User} from "firebase/auth";
 import { getFirestore, collection, addDoc  , getDocs, query, where , deleteDoc , doc, getDoc} from "firebase/firestore";
 
 // Firebase configuration from Firebase Console
@@ -18,4 +18,5 @@ const auth = getAuth();
 setPersistence(auth, browserLocalPersistence);
 const provider = new GoogleAuthProvider();
 
-export { db , collection, addDoc , auth, provider ,getDocs, query, where , deleteDoc , doc , getDoc};
+export { db , collection, addDoc , auth, provider ,getDocs, query, where , deleteDoc , doc , getDoc, getAuth, onAuthStateChanged};
+export type { User };
